Resolve review items
