+++ subprojects/tooling-api/src/main/java/org/gradle/tooling/events/test/TestOutputDescriptor.java
@@ -26,4 +26,14 @@ import org.gradle.tooling.events.OperationDescriptor;
  */
 @Incubating
 public interface TestOutputDescriptor extends OperationDescriptor {
+
+    /**
+     * @return Whether the content was printed to the standard output or the standard error.
+     */
+    Destination getDestination();
+
+    /**
+     * @return The text printed by the test.
+     */
+    String getMessage();
 }
